+++ src/ClientConfig_TEST.cc
@@ -57,11 +57,11 @@ TEST(ClientConfig, ServersCanBeAdded)
 {
   ClientConfig config;
   ServerConfig srv;
-  srv.URL("asdf");
+  srv.URL("http://asdf");
   config.AddServer(srv);
 
   ASSERT_EQ(1u, config.Servers().size());
-  EXPECT_EQ(std::string("asdf"), config.Servers().front().URL());
+  EXPECT_EQ(std::string("http://asdf"), config.Servers().front().URL());
 }
 
 /////////////////////////////////////////////////
@@ -409,6 +409,46 @@ TEST(ClientConfig, AsString)
     EXPECT_NE(str.find("config/path"), std::string::npos);
     EXPECT_NE(str.find("cache/location"), std::string::npos);
     EXPECT_NE(str.find("http://serverurl.com"), std::string::npos);
+  }
+}
+
+/////////////////////////////////////////////////
+TEST(ServerConfig, Url)
+{
+  // Invalid URL string
+  {
+    ServerConfig srv;
+    srv.URL("asdf");
+    EXPECT_TRUE(srv.URL().empty());
+  }
+
+  // Valid URL
+  {
+    ServerConfig srv;
+    srv.URL("http://banana:8080");
+    EXPECT_EQ("http://banana:8080", srv.URL());
+    EXPECT_EQ("http", srv.Url().Scheme());
+    EXPECT_EQ("banana:8080", srv.Url().Path().Str());
+  }
+
+  // Trailing /
+  {
+    ServerConfig srv;
+    srv.URL("http://banana:8080/");
+    EXPECT_EQ("http://banana:8080", srv.URL());
+  }
+
+  // Set from URI
+  {
+    auto url = common::URI();
+    url.SetScheme("http");
+    url.Path() = common::URIPath("banana:8080");
+
+    ServerConfig srv;
+    srv.SetUrl(url);
+    EXPECT_EQ("http://banana:8080", srv.URL());
+    EXPECT_EQ("http", srv.Url().Scheme());
+    EXPECT_EQ("banana:8080", srv.Url().Path().Str());
   }
 }