+++ src/FuelClient.cc
@@ -111,21 +111,21 @@ ClientConfig &FuelClient::Config()
 }
 
 //////////////////////////////////////////////////
-Result FuelClient::ModelDetails(const ServerConfig &_server,
+Result FuelClient::ModelDetails(const ServerConfig &/*_server*/,
   const ModelIdentifier &_id, ModelIdentifier &_model) const
 {
   ignition::fuel_tools::REST rest;
   RESTResponse resp;
 
-  auto serverURL = _server.URL();
-  auto version = _server.Version();
+  auto serverURL = _id.Server().URL();
+  auto version = _id.Server().Version();
   auto path = ignition::common::joinPaths(_id.Owner(), "models", _id.Name());
 
   resp = rest.Request(REST::GET, serverURL, version, path, {}, {}, "");
   if (resp.statusCode != 200)
     return Result(Result::FETCH_ERROR);
 
-  _model = JSONParser::ParseModel(resp.data, _server);
+  _model = JSONParser::ParseModel(resp.data, _id.Server());
 
   return Result(Result::FETCH);
 }
@@ -146,7 +146,7 @@ ModelIter FuelClient::Models(const ServerConfig &_server) const
 }
 
 //////////////////////////////////////////////////
-ModelIter FuelClient::Models(const ServerConfig &_server,
+ModelIter FuelClient::Models(const ServerConfig &/*_server*/,
   const ModelIdentifier &_id) const
 {
   // Check local cache first
@@ -164,7 +164,7 @@ ModelIter FuelClient::Models(const ServerConfig &_server,
   else
     path = ignition::common::joinPaths(_id.Owner(), "models");
 
-  return ModelIterFactory::Create(this->dataPtr->rest, _server, path);
+  return ModelIterFactory::Create(this->dataPtr->rest, _id.Server(), path);
 }
 
 //////////////////////////////////////////////////
@@ -184,14 +184,14 @@ Result FuelClient::DeleteModel(const ServerConfig &/*_server*/,
 }
 
 //////////////////////////////////////////////////
-Result FuelClient::DownloadModel(const ServerConfig &_server,
+Result FuelClient::DownloadModel(const ServerConfig &/*_server*/,
   const ModelIdentifier &_id)
 {
   ignition::fuel_tools::REST rest;
   RESTResponse resp;
 
-  auto serverURL = _server.URL();
-  auto version = _server.Version();
+  auto serverURL = _id.Server().URL();
+  auto version = _id.Server().Version();
   auto path = ignition::common::joinPaths(_id.Owner(), "models",
     _id.Name() + ".zip");
 
@@ -207,7 +207,7 @@ Result FuelClient::DownloadModel(const ServerConfig &_server,
 
 //////////////////////////////////////////////////
 bool FuelClient::ParseModelURL(const std::string &_modelURL,
-  ServerConfig &_srv, ModelIdentifier &_id)
+  ServerConfig &/*_server*/, ModelIdentifier &_id)
 {
   std::smatch match;
   if (!std::regex_match(_modelURL, match, *this->dataPtr->urlModelRegex))
@@ -216,7 +216,7 @@ bool FuelClient::ParseModelURL(const std::string &_modelURL,
   if (match.size() != 6u)
     return false;
 
-  std::string method = match[1];
+  std::string scheme = match[1];
   std::string server = match[2];
   std::string owner = match[4];
   std::string name = match[5];
@@ -224,7 +224,9 @@ bool FuelClient::ParseModelURL(const std::string &_modelURL,
   _id.Owner(owner);
   _id.Name(name);
 
-  _srv.URL(method + "://" + server);
+  ServerConfig serverConfig;
+  serverConfig.URL(scheme + "://" + server);
+  _id.Server(serverConfig);
 
   return true;
 }